Fin de la mission 1

+++ vues/v_listeFraisHorsForfait.php
@@ -15,15 +15,18 @@
         $date = $unFraisHorsForfait['date'];
         $montant = $unFraisHorsForfait['montant'];
         $id = $unFraisHorsForfait['id'];
-        ?>		
-        <tr>
-            <td><?php echo $date ?></td>
-            <td><?php echo $libelle ?></td>
-            <td><?php echo $montant ?></td>
-            <td><a href="index.php?uc=gererFrais&action=supprimerFrais&idFrais=<?php echo $id ?>" 
-                   onclick="return confirm('Voulez-vous vraiment supprimer ce frais?');">Supprimer ce frais</a></td>
-        </tr>
-        <?php
+        $debutLibelle = substr($libelle, 0, 7);
+        if ($debutLibelle != 'REFUSER') {
+            ?>	
+            <tr>
+                <td><?php echo $date ?></td>
+                <td><?php echo $libelle ?></td>
+                <td><?php echo $montant ?></td>
+                <td><a href="index.php?uc=gererFrais&action=supprimerFrais&idFrais=<?php echo $id ?>" 
+                       onclick="return confirm('Voulez-vous vraiment supprimer ce frais?');">Supprimer ce frais</a></td>
+            </tr>
+            <?php
+        }
     }
     ?>
